backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
geoalchemy2
asyncpg
python-dotenv
requests
pandas
pydantic-settings
python-multipart
openai
# addresskr 
